Fix first code review issues
